Enable earth dragging rotation behavior
The rotation control needs to be improved, but it's at least working and
making the global state update.

Part of the idea on how to do this came from here:
http://www.html5gamedevs.com/topic/32800-rotation-control/

And it turns out that the gl-matrix package used there isn't actually
necessary. I was surprised that pixi doesn't seem to have some built-in
mechanism to do this, but maybe that's just how this kind of programming
is... I'm kinda new to interactive graphics stuff like this.