Use Torus instead of lines
This fixes inconsistent line rendering on different systems.

See:
* https://mattdesl.svbtle.com/drawing-lines-is-hard
* https://stackoverflow.com/a/41911520/173630

[THREE.MeshLine](https://github.com/spite/THREE.MeshLine) is one
workaround for this, but for these particular lines I can get by with
just using the Torus construct already included in three.js. This is
better anyways because I've seen some shader errors when using
THREE.MeshLine - it's not quite up to date with upstream three.js.