Add d3
The lightcurve view is really a plot instead of a pixi-like scene. To
handle the axes and scale correctly, I need an SVG plotting library
rather than a Canvas/WebGL drawing library. I've decided d3 is the best way
to go. Although I'm already familiar with jsxgraph (which also renders
in SVG), positioning in jsxgraph can be really non-intuitive... and the
library is really focused towards advanced math visualization that
isn't necessary here. I'm going to have to learn d3, but it's a good
time investment - d3 is pretty much industry standard and not going
anywhere.