Add jsxgraph for plotting scene
I think it's easiest to handle the plot with SVG instead of Canvas
(pixi) because the lines need to retain their width as the view zooms in
and out along the Y axis.

Using JSXGraph will give me the scalable Y axis labels.

So, this is messy right now. Will be cleaned up soon.