implement analemma - closes #83
Well this took a long time. And then it just worked and everything's in
place all of a sudden, after figuring out how to render the original
author's analemma function in Three.js

https://gist.github.com/chris-siedell/b5de8dae41cfa8a5ad67a1501aeeab47#file-analemma-as-L52

Now I need to fix an issue with the sun's movement: its declination
moves it up and down correctly along the analemma, but it needs to move
slightly forward and backward as well over the course of the year to follow
the figure-eight.

You turn on the analemma with the Show Analemma checkbox.