Fix moon positioning bugs
I got rid of the moonObserverPos state: It's confusing and unnecessary.
Only two pieces of data are needed to display the global scene:
* observerAngle (Sun position. This is the rotation of the earth)
* moonAngle (The moon's phase, shown as the orbit circle on MainView)

Note that in the Horizon Diagram, the moon's position isn't the same as
moonAngle. The moon needs to follow along the sun and slowly lag behind
it. That's why I made moonObserverPos in the first place, but it
shouldn't be a global state variable.. instead, it can just be derived
from moonAngle and observerAngle in this scene.