Sun sim: Improve sun dragging behavior
This makes the invisible plane respond to changes in latitude and sun
declination.

The position isn't accurate at high and low latitudes - you can see what
I mean by uncommenting the planeHelper - and this is reflected with
awkward dragging behavior. But, it's working better than it was, and
I've spent a lot of time on this so I'm going to come back to this at
another time.