Port a bunch of actionscript code to JS
From here: https://gist.github.com/chris-siedell/662dd6f5dd253519f172b288520bf537

There's a bunch of code I'm going to port over from Flash in order to
get our calculations correct. There's so much that I'm not certain this
will work, but I'm hoping it will just fall into place like the Sun's
Analemma did.

This is a work in progress, just an intermediate step to get a bunch of
this code in the codebase. I'm not actually using it yet.

And yeah, the code style will need to be updated as well.