Add accurate lightcurve data - #197, #171
This adds accurate eclipse data to the lightcurve view. I still need to
fix the display a bit.

This is really messy, but I think I'm moving forward here. I've gotten a
good amount of this code running in JS, while commenting out a bunch of
operations/functions that aren't necessary (e.g., setting display text
in an imperative way. React can handle this better.)

The next interactive I'll be working on, the Eclipsing Binary simulation,
is probably twice as complicated as this one, so I'll need to get in the
habit of this porting process.