Completely fix lightcurve
There was one secret here that I found by debugging the original code in
flash. I need to set `regionShown` to 'eclipse of body 1', and then
the lightcurve displays correctly in d3, without the need for chopping
off the second half of it like I was doing. I still need to normalize
it and negate the Y values though.