Add dynamic cursor phase
This makes the planet move along the some sliver of orbit as the
lightcurve represents.

A key thing I was missing here is that the actionscript defines a
`cursorPhase` property with a getter and setter, and I need to call
`getCursorPhase()` in order to get the right value, instead of just
looking at the `cursorsPhase` property directly.